pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "metis-client"
dynamic = ["version"]  # read from metis_client/__init__.py
description = """Metis infra API client in Python"""
authors = [{name = "Sergei Korolev", email = "knopki@duck.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: BSD License",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Information Analysis"
]
keywords = ["metis", "client"]
requires-python = ">=3.8"
dependencies = [
    "aiohttp >= 3.7.4",
    "aiohttp-sse-client >= 0.2.1",
    "asgiref >= 3.5.2",
    "typing-extensions >= 4.2.0; python_version < '3.11'",
    "yarl >= 1.6.3",
]

[project.optional-dependencies]
debug = [
      "wdb"
]
lint = [
     "autoflake",
     "black",
     "flake8",
     "flake8-bugbear",
     "isort",
     "mypy >= 1.0.0, <2",
     "pylint",
     "pylint-per-file-ignores >= 1",
     "pyupgrade",
]
test = [
    "freezegun",
    "pytest-aiohttp >= 1.0.4, <2",
    "pytest-cov",
]

[project.urls]
Source = "https://github.com/tilde-lab/metis-client"

[tool.flit.module]
name = "metis_client"

[tool.autoflake]
expand-star-imports = true
ignore-init-module-imports = true
remove-all-unused-imports = true
remove-duplicate-keys = true
remove-unused-variables = true

[tool.black]
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
# see https://github.com/psf/black
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
use_parentheses = true
line_length = 88
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
py_version = 38

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true

[tool.pylint.MASTER]
load-plugins=[
    "pylint_per_file_ignores",
]

[tool.pylint.main]
jobs = 0
py-version = "3.8"
recursive = true
suggestion-mode = true

[tool.pylint.basic]
good-names = ["id", "x", "xy", "y", "_", "__"]

[tool.pylint.design]
max-args = 7
max-attributes = 10
min-public-methods = 1

[tool.pylint.format]
max-line-length = 88

[tool.pylint.reports]
output-format = "colorized"
reports = "no"
score = "no"

[tool.pylint-per-file-ignores]
"/tests/" = "redefined-outer-name"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=metis_client --cov-report=term --cov-report=html --no-cov-on-fail --cov-fail-under=99"
